Update classes to match September 2026 update
m_flLastShootTime and m_bDropped were actually missing, and the first was a float.

+++ game/server/player.h
@@ -1182,7 +1182,7 @@ private:
 
     int				m_nKickErrors;
 
-    float			m_flSpeakTime;
+    double			m_flSpeakTime;
 
     Vector			m_hullMins;
     Vector			m_hullMaxs;
